kafka: move partition|replica_info to partition proxy
This is the main interface that uses the types. Further, the types are
not used in the protocol library, nor the kafka client.
