+++ src/v/kafka/server/partition_proxy.h
@@ -27,6 +27,21 @@
 namespace kafka {
 
 /**
+ * Describes single partition replica. Used by replica selector
+ */
+struct replica_info {
+    model::node_id id;
+    model::offset high_watermark;
+    model::offset log_end_offset;
+    bool is_alive;
+};
+
+struct partition_info {
+    std::vector<replica_info> replicas;
+    std::optional<model::node_id> leader;
+};
+
+/**
  * Kafka layer wrapper for partition, this wrapper performs offsets translation
  * to eliminate offset skew caused by existantece of Raft configuration batches.
  */
